+++ src/redturtle/faq/__init__.py
@@ -3,4 +3,4 @@
 from zope.i18nmessageid import MessageFactory
 
 
-_ = MessageFactory('redturtle.faq')
+_ = MessageFactory("redturtle.faq")