+++ src/redturtle/faq/content.py
@@ -1,17 +1,14 @@
 # -*- coding: utf-8 -*-
-from redturtle.faq.interfaces import IFaq
-from redturtle.faq.interfaces import IFaqFolder
 from plone.dexterity.content import Container
+from redturtle.faq.interfaces import IFaq, IFaqFolder
 from zope.interface import implementer
 
 
 @implementer(IFaq)
 class Faq(Container):
-    """
-    """
+    """ """
 
 
 @implementer(IFaqFolder)
 class FaqFolder(Container):
-    """
-    """
+    """ """