+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Serialize returned blocks.
+  [cekk]
 
 
 1.0.1 (2022-01-17)